chore: move types to one folder
