+++ src/app/api/bitcoin/mempool/route.ts
@@ -1,5 +1,7 @@
-import { NextResponse } from 'next/server';
-import BitcoinCLI, { type MempoolEntry } from '@/lib/bitcoin-cli';
+import { NextResponse } from "next/server";
+
+import BitcoinCLI from "@/lib/bitcoin-cli";
+import { MempoolEntry } from "@/types/bitcoin-cli";
 
 const bitcoin = new BitcoinCLI();
 
@@ -7,12 +9,12 @@ export async function GET() {
   try {
     const [mempoolInfo, rawMempool] = await Promise.all([
       bitcoin.getMempoolInfo(),
-      bitcoin.getRawMempool(true)
+      bitcoin.getRawMempool(true),
     ]);
 
     // Calculate fee rates from mempool entries
     const feeRates = Object.values(rawMempool as Record<string, MempoolEntry>)
-      .map(tx => (tx.fees.base * 100000000) / tx.size) // Convert to sats/vB
+      .map((tx) => (tx.fees.base * 100000000) / tx.size) // Convert to sats/vB
       .sort((a, b) => a - b);
 
     const getFeePercentile = (percentile: number) => {
@@ -23,21 +25,21 @@ export async function GET() {
     return NextResponse.json({
       info: mempoolInfo,
       feeEstimates: {
-        low: getFeePercentile(10),    // 10th percentile
-        medium: getFeePercentile(50),  // 50th percentile
-        high: getFeePercentile(80),    // 80th percentile
-        urgent: getFeePercentile(90),  // 90th percentile
+        low: getFeePercentile(10), // 10th percentile
+        medium: getFeePercentile(50), // 50th percentile
+        high: getFeePercentile(80), // 80th percentile
+        urgent: getFeePercentile(90), // 90th percentile
       },
       transactions: Object.keys(rawMempool).length,
     });
   } catch (error) {
-    console.error('Error fetching mempool:', error);
+    console.error("Error fetching mempool:", error);
     return NextResponse.json(
       {
-        status: 'error',
-        error: 'Failed to fetch mempool data'
+        status: "error",
+        error: "Failed to fetch mempool data",
       },
       { status: 500 }
     );
   }
-} 
+}
